requirements.txt
firebase-admin==5.2.0
google-auth-oauthlib==0.4.6
Pillow==9.0.0
pycountry==22.3.5
pandas==1.4.1
streamlit-authenticator==0.2.1